[CLEANUP] pageType at the cart_form
The original definition in pageType leads to a "TypeError" with plugin.tx_cart.settings.addToCartByAjax = 0.

Possible solution:
1. pageType="{f:if(condition:'{settings.addToCartByAjax}', then:'{settings.addToCartByAjax}', else:'{settings.addToCartByAjax}'')}"
2. pageType="{f:if(condition:'{settings.addToCartByAjax}', then:'{settings.addToCartByAjax}', else:'0'')}"
3. pageType="{settings.addToCartByAjax}"

I decided to use solution 3, because it makes this point clearer. Furthermore with addToCartByAjax the PageType is defined. If you don't want to use the Ajax function you set addToCartByAjax = 0. For this reason no condition is needed at this point.