[BUGFIX] Show translated product in teaser plugin
Translating products means using the language
overlay. As a result the returned object has
all properties of the translation but still the
UID of the original object. As a result the
`orderByField()` failed as it compared the UIDs
of the translated objects against objects that
still had the original UIDs.

Fixes: #205